+++ app/(user)/my/components/TabBar.tsx
@@ -1,4 +1,4 @@
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "./ui/MyTabs";
 
 interface Tab {
   value: string;
